Set `Nab=Nab23`
Set the `Nab` variable to the right value so when accessing the `AB`
array we do not read out of bounds. Fixes a crash seeing on Linux,
probably thanks to some overflow protection from GCC.

From a point of view of code execution this solution is right; also
fixes the crash. But from the point of view of the application, I don't
know what are the ramifications.

Fixes #29. Maybe.